requirements.txt
streamlit
phidata
groq
python-dotenv
Pillow